Screenshot of pipeline pull request
I know it says the tests failed, but I fixed the code to give the expected results for all the test cases used in the testing file. For some reason, the test is failing because of an error where the compiler thinks the functions are not actually functions, but they were indeed declared as function.